flask_app_template/requirements.txt
Flask==2.3.2
flask_apscheduler==1.12.4
Bootstrap-Flask==2.2.0
flask_sqlalchemy==3.0.3
flask_wtf==1.1.1
PyYAML==6.0
SQLAlchemy==2.0.12
WTForms==3.0.1
gunicorn==20.1.0
